Linux: Move some 32-bit ipc types to Types.h
This way we can run the struct verifier over it
