+++ Source/Tests/LinuxSyscalls/x32/Types.h
@@ -9,13 +9,20 @@ $end_info$
 #include <FEXCore/Core/SignalDelegator.h>
 #include <FEXCore/Utils/CompilerDefs.h>
 
+#include <asm/ipcbuf.h>
+#include <asm/shmbuf.h>
 #include <bits/types/stack_t.h>
 #include <cstdint>
 #include <cstring>
 #include <fcntl.h>
+#include <limits>
+#include <sys/ipc.h>
 #include <mqueue.h>
 #include <signal.h>
+#include <sys/msg.h>
 #include <sys/resource.h>
+#include <sys/sem.h>
+#include <sys/shm.h>
 #include <sys/socket.h>
 #include <sys/stat.h>
 #include <sys/statfs.h>
@@ -934,10 +941,10 @@ rlimit32 {
     val.rlim_max = rlim_max;
 
     if (val.rlim_cur == ~0U) {
-      val.rlim_cur = ~0ULL;
+      val.rlim_cur = ~0UL;
     }
     if (val.rlim_max == ~0U) {
-      val.rlim_max = ~0ULL;
+      val.rlim_max = ~0UL;
     }
 
     return val;
@@ -1191,5 +1198,445 @@ epoll_event32 {
 };
 static_assert(std::is_trivial<epoll_event32>::value, "Needs to be trivial");
 static_assert(sizeof(epoll_event32) == 12, "Incorrect size");
+
+struct ipc_perm_32 {
+  uint32_t key;
+  uint16_t uid;
+  uint16_t gid;
+  uint16_t cuid;
+  uint16_t cgid;
+  uint16_t mode;
+  uint16_t seq;
+
+  ipc_perm_32() = delete;
+
+  operator struct ipc_perm() const {
+    struct ipc_perm perm;
+    perm.__key = key;
+    perm.uid   = uid;
+    perm.gid   = gid;
+    perm.cuid  = cuid;
+    perm.cgid  = cgid;
+    perm.mode  = mode;
+    perm.__seq = seq;
+    return perm;
+  }
+
+  ipc_perm_32(struct ipc_perm perm) {
+    key  = perm.__key;
+    uid  = perm.uid;
+    gid  = perm.gid;
+    cuid = perm.cuid;
+    cgid = perm.cgid;
+    mode = perm.mode;
+    seq  = perm.__seq;
+  }
+};
+
+static_assert(std::is_trivial<ipc_perm_32>::value, "Needs to be trivial");
+static_assert(sizeof(ipc_perm_32) == 16, "Incorrect size");
+
+struct ipc_perm_64 {
+  uint32_t key;
+  uint32_t uid;
+  uint32_t gid;
+  uint32_t cuid;
+  uint32_t cgid;
+  uint16_t mode;
+  uint16_t _pad1;
+  uint16_t seq;
+  uint16_t _pad2;
+  compat_ulong_t _pad[2];
+
+  ipc_perm_64() = delete;
+
+  operator struct ipc_perm() const {
+    struct ipc_perm perm;
+    perm.__key = key;
+    perm.uid   = uid;
+    perm.gid   = gid;
+    perm.cuid  = cuid;
+    perm.cgid  = cgid;
+    perm.mode  = mode;
+    perm.__seq = seq;
+    return perm;
+  }
+
+  ipc_perm_64(struct ipc_perm perm) {
+    key  = perm.__key;
+    uid  = perm.uid;
+    gid  = perm.gid;
+    cuid = perm.cuid;
+    cgid = perm.cgid;
+    mode = perm.mode;
+    seq  = perm.__seq;
+  }
+};
+
+static_assert(std::is_trivial<ipc_perm_64>::value, "Needs to be trivial");
+static_assert(sizeof(ipc_perm_64) == 36, "Incorrect size");
+
+struct shmid_ds_32 {
+  ipc_perm_32 shm_perm;
+  int32_t shm_segsz;
+  int32_t shm_atime;
+  int32_t shm_dtime;
+  int32_t shm_ctime;
+  uint16_t shm_cpid;
+  uint16_t shm_lpid;
+  uint16_t shm_nattch;
+  uint16_t shm_unused;
+  uint32_t shm_unused2;
+  uint32_t shm_unused3;
+
+  shmid_ds_32() = delete;
+
+  operator struct shmid_ds() const {
+    struct shmid_ds buf;
+    buf.shm_perm = shm_perm;
+
+    buf.shm_segsz = shm_segsz;
+    buf.shm_atime = shm_atime;
+    buf.shm_dtime = shm_dtime;
+    buf.shm_ctime = shm_ctime;
+    buf.shm_cpid = shm_cpid;
+    buf.shm_lpid = shm_lpid;
+    buf.shm_nattch = shm_nattch;
+    return buf;
+  }
+
+  shmid_ds_32(struct shmid_ds buf)
+    : shm_perm {buf.shm_perm} {
+    shm_segsz = buf.shm_segsz;
+    shm_atime = buf.shm_atime;
+    shm_dtime = buf.shm_dtime;
+    shm_ctime = buf.shm_ctime;
+    shm_cpid = buf.shm_cpid;
+    shm_lpid = buf.shm_lpid;
+    shm_nattch = buf.shm_nattch;
+  }
+};
+
+static_assert(std::is_trivial<shmid_ds_32>::value, "Needs to be trivial");
+static_assert(sizeof(shmid_ds_32) == 48, "Incorrect size");
+
+struct shmid_ds_64 {
+  ipc_perm_64 shm_perm;
+  compat_size_t shm_segsz;
+  compat_ulong_t shm_atime;
+  compat_ulong_t shm_atime_high;
+  compat_ulong_t shm_dtime;
+  compat_ulong_t shm_dtime_high;
+  compat_ulong_t shm_ctime;
+  compat_ulong_t shm_ctime_high;
+  int32_t shm_cpid;
+  int32_t shm_lpid;
+  compat_ulong_t shm_nattch;
+  compat_ulong_t shm_unused4;
+  compat_ulong_t shm_unused5;
+
+  shmid_ds_64() = delete;
+
+  operator struct shmid_ds() const {
+    struct shmid_ds buf;
+    buf.shm_perm = shm_perm;
+
+    buf.shm_segsz = shm_segsz;
+    buf.shm_atime = shm_atime_high;
+    buf.shm_atime <<= 32;
+    buf.shm_atime |= shm_atime;
+
+    buf.shm_dtime = shm_dtime_high;
+    buf.shm_dtime <<= 32;
+    buf.shm_dtime |= shm_dtime;
+
+    buf.shm_ctime = shm_ctime_high;
+    buf.shm_ctime <<= 32;
+    buf.shm_ctime |= shm_ctime;
+
+    buf.shm_cpid = shm_cpid;
+    buf.shm_lpid = shm_lpid;
+    buf.shm_nattch = shm_nattch;
+    return buf;
+  }
+
+  shmid_ds_64(struct shmid_ds buf)
+    : shm_perm {buf.shm_perm} {
+    shm_segsz = buf.shm_segsz;
+    shm_atime = buf.shm_atime;
+    shm_atime_high = buf.shm_atime >> 32;
+    shm_dtime = buf.shm_dtime;
+    shm_dtime_high = buf.shm_dtime >> 32;
+    shm_ctime = buf.shm_ctime;
+    shm_ctime_high = buf.shm_ctime >> 32;
+    shm_cpid = buf.shm_cpid;
+    shm_lpid = buf.shm_lpid;
+    shm_nattch = buf.shm_nattch;
+  }
+};
+
+static_assert(std::is_trivial<shmid_ds_64>::value, "Needs to be trivial");
+static_assert(sizeof(shmid_ds_64) == 84, "Incorrect size");
+
+struct semid_ds_32 {
+  struct ipc_perm_32 sem_perm;
+  int32_t sem_otime;
+  int32_t sem_ctime;
+  uint32_t sem_base;
+  uint32_t sem_pending;
+  uint32_t sem_pending_last;
+  uint32_t undo;
+  uint16_t sem_nsems;
+  uint16_t _pad;
+
+  semid_ds_32() = delete;
+
+  operator struct semid_ds() const {
+    struct semid_ds buf{};
+    buf.sem_perm = sem_perm;
+
+    buf.sem_otime = sem_otime;
+    buf.sem_ctime = sem_ctime;
+    buf.sem_nsems = sem_nsems;
+
+    // sem_base, sem_pending, sem_pending_last, undo doesn't exist in the definition
+    // Kernel doesn't return anything in them
+    return buf;
+  }
+
+  semid_ds_32(struct semid_ds buf)
+    : sem_perm {buf.sem_perm} {
+    sem_otime = buf.sem_otime;
+    sem_ctime = buf.sem_ctime;
+    sem_nsems = buf.sem_nsems;
+  }
+};
+
+static_assert(std::is_trivial<semid_ds_32>::value, "Needs to be trivial");
+static_assert(sizeof(semid_ds_32) == 44, "Incorrect size");
+
+struct semid_ds_64 {
+  struct ipc_perm_64 sem_perm;
+  uint32_t sem_otime;
+  uint32_t sem_otime_high;
+  uint32_t sem_ctime;
+  uint32_t sem_ctime_high;
+  uint32_t sem_nsems;
+  uint32_t _pad[2];
+
+  semid_ds_64() = delete;
+
+  operator struct semid_ds() const {
+    struct semid_ds buf{};
+    buf.sem_perm = sem_perm;
+
+    buf.sem_otime = sem_otime_high;
+    buf.sem_otime <<= 32;
+    buf.sem_otime |= sem_otime;
+    buf.sem_ctime = sem_ctime_high;
+    buf.sem_ctime <<= 32;
+    buf.sem_ctime |= sem_ctime;
+    buf.sem_nsems = sem_nsems;
+
+    // sem_base, sem_pending, sem_pending_last, undo doesn't exist in the definition
+    // Kernel doesn't return anything in them
+    return buf;
+  }
+
+  semid_ds_64(struct semid_ds buf)
+    : sem_perm {buf.sem_perm} {
+    sem_otime = buf.sem_otime;
+    sem_otime_high = buf.sem_otime >> 32;
+    sem_ctime = buf.sem_ctime;
+    sem_ctime_high = buf.sem_ctime >> 32;
+    sem_nsems = buf.sem_nsems;
+  }
+};
+
+static_assert(std::is_trivial<semid_ds_64>::value, "Needs to be trivial");
+static_assert(sizeof(semid_ds_64) == 64, "Incorrect size");
+
+struct msqid_ds_32 {
+  struct ipc_perm_32 msg_perm;
+  compat_uptr_t msg_first;
+  compat_uptr_t msg_last;
+  uint32_t msg_stime;
+  uint32_t msg_rtime;
+  uint32_t msg_ctime;
+  uint32_t msg_lcbytes;
+  uint32_t msg_lqbytes;
+  uint16_t msg_cbytes;
+  uint16_t msg_qnum;
+  uint16_t msg_qbytes;
+  uint16_t msg_lspid;
+  uint16_t msg_lrpid;
+
+  msqid_ds_32() = delete;
+  msqid_ds_32(struct msqid_ds buf)
+    : msg_perm {buf.msg_perm} {
+    // msg_first and msg_last are unused and untouched
+    msg_stime = buf.msg_stime;
+    msg_rtime = buf.msg_rtime;
+    msg_ctime = buf.msg_ctime;
+    if (buf.msg_cbytes > std::numeric_limits<uint16_t>::max()) {
+      msg_cbytes = std::numeric_limits<uint16_t>::max();
+    }
+    else {
+      msg_cbytes = buf.msg_cbytes;
+    }
+    msg_lcbytes = buf.msg_cbytes;
+
+    if (buf.msg_qnum > std::numeric_limits<uint16_t>::max()) {
+      msg_qnum = std::numeric_limits<uint16_t>::max();
+    }
+    else {
+      msg_qnum = buf.msg_qnum;
+    }
+
+    if (buf.msg_cbytes > std::numeric_limits<uint16_t>::max()) {
+      msg_cbytes = std::numeric_limits<uint16_t>::max();
+    }
+    else {
+      msg_cbytes = buf.msg_cbytes;
+    }
+    msg_lqbytes = buf.msg_qbytes;
+    msg_lspid = buf.msg_lspid;
+    msg_lrpid = buf.msg_lrpid;
+  }
+};
+static_assert(std::is_trivial<msqid_ds_32>::value, "Needs to be trivial");
+static_assert(sizeof(msqid_ds_32) == 56, "Incorrect size");
+
+struct msqid_ds_64 {
+  struct ipc_perm_64 msg_perm;
+  uint32_t msg_stime;
+  uint32_t msg_stime_high;
+  uint32_t msg_rtime;
+  uint32_t msg_rtime_high;
+  uint32_t msg_ctime;
+  uint32_t msg_ctime_high;
+  uint32_t msg_cbytes;
+  uint32_t msg_qnum;
+  uint32_t msg_qbytes;
+  uint32_t msg_lspid;
+  uint32_t msg_lrpid;
+  uint32_t _pad[2];
+
+  msqid_ds_64() = delete;
+  msqid_ds_64(struct msqid_ds buf)
+    : msg_perm {buf.msg_perm} {
+    msg_stime = buf.msg_stime;
+    msg_stime_high = buf.msg_stime >> 32;
+    msg_rtime = buf.msg_rtime;
+    msg_rtime_high = buf.msg_rtime >> 32;
+    msg_ctime = buf.msg_ctime;
+    msg_ctime_high = buf.msg_ctime >> 32;
+    msg_cbytes = buf.msg_cbytes;
+    msg_qnum = buf.msg_qnum;
+    msg_qbytes = buf.msg_qbytes;
+    msg_lspid = buf.msg_lspid;
+    msg_lrpid = buf.msg_lrpid;
+  }
+};
+
+static_assert(std::is_trivial<msqid_ds_64>::value, "Needs to be trivial");
+static_assert(sizeof(msqid_ds_64) == 88, "Incorrect size");
+
+struct
+FEX_ANNOTATE("fex-match")
+shminfo_32 {
+  uint32_t shmmax;
+  uint32_t shmmin;
+  uint32_t shmmni;
+  uint32_t shmseg;
+  uint32_t shmall;
+
+  shminfo_32() = delete;
+
+  operator struct shminfo() const {
+    struct shminfo si;
+    si.shmmax = shmmax;
+    si.shmmin = shmmin;
+    si.shmmni = shmmni;
+    si.shmseg = shmseg;
+    si.shmall = shmall;
+    return si;
+  }
+
+  shminfo_32(struct shminfo si) {
+    shmmax = si.shmmax;
+    shmmin = si.shmmin;
+    shmmni = si.shmmni;
+    shmseg = si.shmseg;
+    shmall = si.shmall;
+  }
+};
+
+static_assert(std::is_trivial<shminfo_32>::value, "Needs to be trivial");
+static_assert(sizeof(shminfo_32) == 20, "Incorrect size");
+
+struct
+FEX_ANNOTATE("alias-x86_32-shminfo64")
+FEX_ANNOTATE("fex-match")
+shminfo_64 {
+  compat_ulong_t shmmax;
+  compat_ulong_t shmmin;
+  compat_ulong_t shmmni;
+  compat_ulong_t shmseg;
+  compat_ulong_t shmall;
+  compat_ulong_t __unused1;
+  compat_ulong_t __unused2;
+  compat_ulong_t __unused3;
+  compat_ulong_t __unused4;
+
+  shminfo_64() = delete;
+
+  operator struct shminfo() const {
+    struct shminfo si;
+    si.shmmax = shmmax;
+    si.shmmin = shmmin;
+    si.shmmni = shmmni;
+    si.shmseg = shmseg;
+    si.shmall = shmall;
+    return si;
+  }
+
+  shminfo_64(struct shminfo si) {
+    shmmax = si.shmmax;
+    shmmin = si.shmmin;
+    shmmni = si.shmmni;
+    shmseg = si.shmseg;
+    shmall = si.shmall;
+  }
+};
+
+static_assert(std::is_trivial<shminfo_64>::value, "Needs to be trivial");
+static_assert(sizeof(shminfo_64) == 36, "Incorrect size");
+
+struct
+FEX_ANNOTATE("alias-x86_32-shm_info")
+FEX_ANNOTATE("fex-match")
+shm_info_32 {
+  int used_ids;
+  uint32_t shm_tot;
+  uint32_t shm_rss;
+  uint32_t shm_swp;
+  uint32_t swap_attempts;
+  uint32_t swap_successes;
+
+  shm_info_32() = delete;
+
+  shm_info_32(struct shm_info si) {
+    used_ids = si.used_ids;
+    shm_tot = si.shm_tot;
+    shm_rss = si.shm_rss;
+    shm_swp = si.shm_swp;
+    swap_attempts = si.swap_attempts;
+    swap_successes = si.swap_successes;
+  }
+};
+
+static_assert(std::is_trivial<shm_info_32>::value, "Needs to be trivial");
+static_assert(sizeof(shm_info_32) == 24, "Incorrect size");
 
 }
